+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/datagen/datapack/ModDimensions.kt
@@ -4,6 +4,7 @@ import dev.aaronhowser.mods.irregular_implements.util.OtherUtil
 import net.minecraft.core.registries.Registries
 import net.minecraft.data.worldgen.BootstrapContext
 import net.minecraft.resources.ResourceKey
+import net.minecraft.resources.ResourceLocation
 import net.minecraft.tags.BlockTags
 import net.minecraft.util.valueproviders.ConstantInt
 import net.minecraft.world.level.Level
@@ -13,25 +14,24 @@ import net.minecraft.world.level.levelgen.FlatLevelSource
 import net.minecraft.world.level.levelgen.flat.FlatLevelGeneratorSettings
 import java.util.*
 
-
 object ModDimensions {
 
-	val SPECTRE_RL = OtherUtil.modResource("spectre")
+	val SPECTRE_ID: ResourceLocation = OtherUtil.modResource("spectre")
 
-	val SPECTRE_STEM_KEY: ResourceKey<LevelStem> =
-		ResourceKey.create(Registries.LEVEL_STEM, SPECTRE_RL)
+	val SPECTRE_LEVEL_STEM_KEY: ResourceKey<LevelStem> =
+		ResourceKey.create(Registries.LEVEL_STEM, SPECTRE_ID)
 
 	val SPECTRE_LEVEL_KEY: ResourceKey<Level> =
-		ResourceKey.create(Registries.DIMENSION, SPECTRE_RL)
+		ResourceKey.create(Registries.DIMENSION, SPECTRE_ID)
 
-	val SPECTRE_DIM_TYPE_KEY: ResourceKey<DimensionType> =
-		ResourceKey.create(Registries.DIMENSION_TYPE, SPECTRE_RL)
+	val SPECTRE_DIMENSION_TYPE_KEY: ResourceKey<DimensionType> =
+		ResourceKey.create(Registries.DIMENSION_TYPE, SPECTRE_ID)
 
-	fun bootstrapType(context: BootstrapContext<DimensionType>) {
+	fun bootstrapDimensionTypes(context: BootstrapContext<DimensionType>) {
 		context.register(
-			SPECTRE_DIM_TYPE_KEY,
+			SPECTRE_DIMENSION_TYPE_KEY,
 			DimensionType(
-				OptionalLong.of(12000L),
+				OptionalLong.of(6000L),
 				false,
 				false,
 				false,
@@ -43,31 +43,31 @@ object ModDimensions {
 				256,
 				256,
 				BlockTags.INFINIBURN_OVERWORLD,
-				SPECTRE_RL,
+				SPECTRE_ID,
 				1f,
 				DimensionType.MonsterSettings(false, false, ConstantInt.of(0), 0)
 			)
 		)
 	}
 
-	fun bootstrapStem(context: BootstrapContext<LevelStem>) {
-		val dimTypeRegistry = context.lookup(Registries.DIMENSION_TYPE)
+	fun bootstrapLevelStems(context: BootstrapContext<LevelStem>) {
+		val dimensionTypeRegistry = context.lookup(Registries.DIMENSION_TYPE)
 		val biomeRegistry = context.lookup(Registries.BIOME)
 		val structureSetRegistry = context.lookup(Registries.STRUCTURE_SET)
 		val placedFeatureRegistry = context.lookup(Registries.PLACED_FEATURE)
 
-		val flatLevelGenSettings =
+		val generatorSettings =
 			FlatLevelGeneratorSettings.getDefault(biomeRegistry, structureSetRegistry, placedFeatureRegistry)
 				.withBiomeAndLayers(
 					listOf(),
 					Optional.empty(),
-					biomeRegistry.get(ModBiomes.SPECTRAL_BIOME_RK).get()
+					biomeRegistry.getOrThrow(ModBiomes.SPECTRAL_BIOME_RK)
 				)
 
-		val flatLevelSource = FlatLevelSource(flatLevelGenSettings)
+		val generator = FlatLevelSource(generatorSettings)
+		val dimensionType = dimensionTypeRegistry.getOrThrow(SPECTRE_DIMENSION_TYPE_KEY)
 
-		val stem = LevelStem(dimTypeRegistry.getOrThrow(SPECTRE_DIM_TYPE_KEY), flatLevelSource)
-		context.register(SPECTRE_STEM_KEY, stem)
+		context.register(SPECTRE_LEVEL_STEM_KEY, LevelStem(dimensionType, generator))
 	}
 
 }